feat : ai 챗봇  아키텍쳐 1차 설계
