+++ apps/recommender/services/chatbot/data/spot_id_map.json
