+++ apps/recommender/services/chatbot/data/spot_index.faiss
