+++ apps/recommender/services/chatbot/data/spot_metadata.json
