+++ apps/recommender/services/chatbot/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from chatbot.views import ChatbotAsyncView
+
+urlpatterns = [
+    path("chat/", ChatbotAsyncView.as_view(), name="chatbot_api"),
+]
